pyproject.toml
[tool.poetry]
name = "nothing-cli"
version = "0.1.0"
description = "Nothing helps coder be more smarter & less dumber."
authors = ["Ainsley McGrath <mcgrath.ainsley@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.scripts]
not = "nothing_cli.main:app"

[tool.poetry.dependencies]
python = "^3.7"
colorama = "^0.4.3"
typer = "^0.1.0"
pydantic = "^1.4"
"ruamel.yaml" = "^0.16.10"
typing-extensions = "^3.7.4"
python-slugify = "^4.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
flake8 = "^3.7.9"
black = "^19.10b0"
ipython = "^7.13.0"
pytest = "^5.4.1"
pre-commit = "^2.5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
